pyproject.toml
[project]
name = "trino-llm-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "wchoi",email = "wchoi@lycorp.co.jp"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "mongoengine (>=0.29.1,<0.30.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "zenml[server] (==0.85.0)",
    "beautifulsoup4 (>=4.13.5,<5.0.0)",
    "chromedriver-autoinstaller (>=0.6.4,<0.7.0)",
    "selenium (>=4.35.0,<5.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "qdrant-client (>=1.15.1,<2.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "sentence-transformers (>=5.1.1,<6.0.0)"
]

[tool.poetry]
packages = [{include = "trino_llm_agent", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "~3.11"
pymongo = "^4.6.2"
click = "^8.0.1"
rich = "^13.7.1"
numpy = "^1.26.4"
poethepoet = "0.29.0"
datasets = "^3.0.1"
torch = "2.2.2"

# Digital data ETL
webdriver-manager = "^4.0.1"
html2text = "^2024.2.26"
jmespath = "^1.0.1"

# Feature engineering
langchain = "^0.2.11"

# RAG
langchain-openai = "^0.1.3"
jinja2 = "^3.1.4"
tiktoken = "^0.7.0"
fake-useragent = "^1.5.1"
langchain-community = "^0.2.11"

# Inference
fastapi = ">=0.100,<=0.110"
uvicorn = "^0.30.6"
opik = "^0.2.2"


[tool.poe.tasks]
local-docker-infrastructure-up = "docker compose up -d"
local-docker-infrastructure-down = "docker compose stop"

local-zenml-server-down = "poetry run zenml logout --local"

local-infrastructure-up = [
    "local-docker-infrastructure-up",
    "local-zenml-server-up"
]

local-infrastructure-down = [
    "local-docker-infrastructure-down",
    "local-zenml-server-down"
]

[tool.poe.tasks.local-zenml-server-up]
env = { OBJC_DISABLE_INITIALIZE_FORK_SAFETY = "YES" }
cmd = "poetry run zenml login --local"

